telosmaindex - telosbookdex splitted in two contracts
