+++ contracts/telosmaindex/CMakeLists.txt
@@ -1,0 +1,10 @@
+cmake_minimum_required(VERSION 3.5)
+project(telos_dex VERSION 0.1.0)
+set(EOSIO_WASM_OLD_BEHAVIOR "Off")
+find_package(eosio.cdt)
+include(EosioWasmToolchain)
+
+add_contract(telosmaindex telosmaindex telosmaindex.cpp)
+
+file(GLOB custom_includes ${CUSTOM_INCLUDES_DIR}/*)
+target_include_directories(telosmaindex PUBLIC ${custom_includes})
