+++ include/vapaee/book/constants.hpp
@@ -1,0 +1,7 @@
+// aux_generate_sell_order
+#define TEXT_AGSO_1 "future payment for order"
+
+// aux_cancel_sell_order
+#define TEXT_ACSO_1 "order canceled, payment returned"
+#define TEXT_ACSO_2 "order canceled, payment returned, because of maintenance"
+
