+++ include/vapaee/book/tables.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+namespace vapaee {
+    namespace book {
+        
+        #include <vapaee/book/tables.all.hpp>
+
+    };
+};
+
