+++ include/vapaee/dex/modules/utils.hpp
@@ -94,17 +94,6 @@ namespace vapaee {
                 PRINT("vapaee::dex::utils::aux_get_real_asset() ...\n");
                 return real;
             }
-            
-            // taken from here
-            // https://github.com/EOSIO/eosio.cdt/issues/404#issuecomment-461626338
-            float stof(std::string s, float def) {
-                if (s == "") return def;
-                std::size_t i = s.find(".");
-                int digits = s.length() - i - 1;
-                s.erase(i, 1); 
-                float result = (float) ( atoi(s.c_str()) / pow(10, digits) );
-                return result;
-            }  
 
         };       
     };
