+++ include/vapaee/dex/tables.all.hpp
@@ -2,23 +2,19 @@
 #include "./tables/blacklist.hpp"
 #include "./tables/historyblock.hpp"
 #include "./tables/delmarkets.hpp"
-#include "./tables/deposits.hpp"
-#include "./tables/depusers.hpp"
 #include "./tables/events.hpp"
 #include "./tables/exp.hpp"
 #include "./tables/history.hpp"
 #include "./tables/historyall.hpp"
 #include "./tables/clients.hpp"
 #include "./tables/markets.hpp"
-#include "./tables/ordersummary.hpp"
 #include "./tables/points.hpp"
-#include "./tables/sellorders.hpp"
 #include "./tables/stat.hpp"
 #include "./tables/state.hpp"
 #include "./tables/last24hs.hpp"
 #include "./tables/tokendata.hpp"
-#include "./tables/tokenevents.hpp"
 #include "./tables/tokengroups.hpp"
 #include "./tables/tokens.hpp"
-#include "./tables/userorders.hpp"
 #include "./tables/whitelist.hpp"
+
+#include "<vapaee/book/tables/deposits.hpp>"
